# docs/swagger class-07.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "defaultDescription",
    "version" : "0.1",
    "title" : "defaultTitle"
  },
  "paths" : {
    "/products/2" : {
      "put" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model4"
          },
          "x-examples" : {
            "application/json" : "{\r\n    \"category\": \"name\",\r\n      \"name\": \"hisham\",\r\n      \"display_name\": \"developer\",\r\n      \"description\": \"Full stack web developer at ASAC\"\r\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model5"
            }
          }
        }
      }
    },
    "/categories" : {
      "get" : {
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model6"
            }
          }
        }
      },
      "post" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model2"
          },
          "x-examples" : {
            "application/json" : "{\r\n    \"name\": \"Hisham new 1111\",\r\n      \"display_name\": \"civil\",\r\n      \"description\": \"from yarmouk\"\r\n}"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/categories/1" : {
      "get" : {
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model7"
            }
          }
        }
      },
      "put" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model0"
          },
          "x-examples" : {
            "application/json" : "{\r\n    \"name\": \"Hisham hello 1\",\r\n      \"display_name\": \"civil\",\r\n      \"description\": \"from yarmouk\"\r\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model8"
            }
          }
        }
      }
    },
    "/categories/{param0}" : {
      "delete" : {
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model1"
          },
          "x-examples" : {
            "application/json" : "{\r\n    \"name\": \"Hisham hello 1\",\r\n      \"display_name\": \"civil\",\r\n      \"description\": \"from yarmouk\"\r\n}"
          }
        }, {
          "name" : "param0",
          "in" : "path",
          "description" : "Example values form path are: '1' and '2'",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "schema" : {
              "$ref" : "#/definitions/Model10"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Model0" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "Model1" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "Model2" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "Model4" : {
      "properties" : {
        "category" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "Model5" : {
      "properties" : {
        "category" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Results" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Model6" : {
      "properties" : {
        "count" : {
          "type" : "integer",
          "format" : "int32"
        },
        "results" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Results"
          }
        }
      }
    },
    "Model7" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Model8" : {
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "display_name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "idToUpdate" : {
          "type" : "string"
        }
      }
    },
    "Model10" : { }
  }
}